moving native language handlers to LibNative
